+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 //import logo from './logo.svg';
 import './App.css';
-import Table from './components/Table'
+import Table from './components/Table/Table'
 
 class App extends React.Component {
   render () {
